Card components broken out and layout fixed
